Refactor geomag code to avoid redundant file I/O
Move getshc() calls out of get_field_components() and into
read_model() so that we are only reading the contents of the coefficients
file once. This means saving all of the coefficients (the gh1 and gh2
arrays) in the BFieldModel struct. This means that we only need to open
(and close) the FILE stream once, so now we pass the stream to getshc()
directly, instead of it opening and closing a stream and being passed
a filename.

All of these changes also mean that get_field_components() now does not
need to take a filename.

The gha and ghb arrays are no longer global, and are moved into
get_field_components. These arrays are the interpolated or extrapolated
spherical harmonic coefficients computed from the gh1 and gh2 arrays,
which are the components obtained from the model.